* release 0.0.3

+++ index.js
@@ -3,7 +3,8 @@
 'use strict';
 
 var 
-    lodash = require('lodash'),
+    _ = require('lodash'),
+    fsJson = require('fs-json')(),
     async = require('async'),
     github = require('octonode'),
     client = github.client(),
@@ -14,20 +15,21 @@ var
     clone = require('nodegit').Clone.clone,
     fs = require('fs'),
     rimraf = require('rimraf'),
-    cancelOption = 'cancel...';
+    cancelOption = '[cancel]',
+    projectEntriesPath = '.projects.json';
 
 program
     .version('0.0.1');
     
 program    
-    .command('list')
+    .command('install')
     .description('List installable projects.')
     .action(list);
     
-program    
-    .command('install [project]')
-    .description('Install an Operation Spark project into your website workspace.')
-    .action(installProject);
+// program    
+//     .command('install [project]')
+//     .description('Install an Operation Spark project into your website workspace.')
+//     .action(installProject);
     
 program.parse(process.argv);
 
@@ -36,58 +38,74 @@ function list() {
     var names = [];
     
     opspark.repos(1, 100, function (err, repos) {
-        if (err) return console.log(err.red);
+        if (err) return console.log(err + ''.red);
         var projects = repos.filter(function (repo) {
             //console.log(repo.description.blue);
             return /PROJECT::/.test(repo.description);
         });
         
-        projects.forEach(function (repo) {
-            names.push(repo.name);
-        });
-        
-        promptForInstall(names);
+        promptForInstall(projects);
     });
 }
 
-function promptForInstall(names) {
-    var promptFor = [{
-        type: "list",
-        name: "project",
-        message: "Select the project you wish to install",
-        choices: names.concat(cancelOption)
-    }];
-    
-    inquirer.prompt(promptFor, function(response) {
-        if(response.project === cancelOption) {
-            console.log('Installtion cancelled, bye bye!'.green);
-            process.exit();
+function promptForInstall(projects) {
+    async.waterfall([
+        function(next) {
+            inquirer.prompt([{
+                type: "list",
+                name: "project",
+                message: "Select the project you wish to install",
+                choices: _.pluck(projects, 'name').concat(cancelOption)}], 
+                function(response) {
+                    if(response.project === cancelOption) {
+                        console.log('Installtion cancelled, bye bye!'.green);
+                        process.exit();
+                    }
+                    next(null, _.where(projects, {'name': response.project})[0]);
+            });
+        },
+        function(project, next) {
+            inquirer.prompt([{
+                        type: "confirm",
+                        name: "install",
+                        message: "You selected " + project.name + ": Go ahead and install?",
+                        default: true
+                }],
+                function(confirm) {
+                    if (confirm.install) { installProject(project); }
+                });
         }
-        installProject(response.project);
-    });
+    ]);
 }
 
 function installProject(project) {
-    if (fs.existsSync(__dirname + '/' + project)) return console.log('Project %s already installed! Please delete manually before reinstalling, or install another project.', project);
+    // TODO : Provide overloadeded method to install with name //
+    //if (!projectName) return list();
+    var projectName = project.name;
     
-    console.log('Installing project %s... please wait...'.green, project);
-    var uri = 'https://github.com/OperationSpark/' + project;
+    if (fs.existsSync(__dirname + '/' + projectName)) return console.log('Project %s already installed! Please delete manually before reinstalling, or install another project.', projectName);
     
-    clone(uri, project, null)
+    console.log('Installing project %s... please wait...'.green, projectName);
+    var uri = 'https://github.com/OperationSpark/' + projectName;
+    
+    clone(uri, projectName, null)
         .then(function(repo) {
             console.log('Successfully cloned project!'.green);
             async.series(
                 [
                     function (next) {
-                        removeGitRemnants(project, next);
+                        removeGitRemnants(projectName, next);
                     },
                     function (next) {
-                        installBower(project, next);    
+                        installBower(projectName, next);
+                    },
+                    function (next) {
+                        appendProjectEntry(project, next);
                     }
                 ],
                 function(err, result){
-                    if (err) return console.log(err.red);
-                    console.log('Installation of project %s complete! Have fun!!!'.blue, project);
+                    if (err) return console.log(err + ''.red);
+                    console.log('Installation of project %s complete! Have fun!!!'.blue, projectName);
                 }
             );
         }, function (err) {
@@ -95,7 +113,21 @@ function installProject(project) {
         });
 }
 
+function appendProjectEntry(project, callback) {
+    var projectEntries = fsJson.loadSync(projectEntriesPath);
+    projectEntries = (projectEntries ? projectEntries : {projects: []});
+    projectEntries.projects.push({
+        name: project.name, 
+        description: project.description.replace('PROJECT:: ', ''), 
+        date: new Date().toLocaleDateString('en-US', { weekday: 'long', year: 'numeric', month: 'long', day: 'numeric'})
+    });
+    fsJson.saveSync(projectEntriesPath, projectEntries);
+    console.log('Project entry saved!'.green);
+    callback();
+}
+
 function installBower(project, callback) {
+    if (!fs.existsSync(__dirname + '/' + project + '/bower.json')) return callback();
     console.log('Installing bower components, please wait...'.green);
     var exec = require('child_process').exec;
     var child = exec('cd ' + project + ' && bower install -F', function(err, stdout, stderr) {
